requirements.txt
torch
numpy
pandas
scikit-learn
umap-learn
matplotlib
networkx
